+++ CotEditor/Sources/CETextView.h
@@ -123,6 +123,14 @@
 @end
 
 
+@interface CETextView (Indentation)
+
+- (IBAction)convertIndentationToSpaces:(nullable id)sender;
+- (IBAction)convertIndentationToTabs:(nullable id)sender;
+
+@end
+
+
 @interface CETextView (LineProcessing)
 
 - (IBAction)moveLineUp:(nullable id)sender;